Extracting logic for parsing for better UT coverage
