+++ src/storage/v2/CMakeLists.txt
@@ -12,8 +12,7 @@ set(storage_v2_src_files
     vertex_accessor.cpp
     storage.cpp)
 
-##### Replication #####
-
+# #### Replication #####
 define_add_lcp(add_lcp_storage lcp_storage_cpp_files generated_lcp_storage_files)
 
 add_lcp_storage(replication/rpc.lcp SLK_SERIALIZE)
@@ -26,10 +25,10 @@ set(storage_v2_src_files
     replication/replication_server.cpp
     replication/serialization.cpp
     replication/slk.cpp
+    replication/replication_persistance_helper.cpp
     ${lcp_storage_cpp_files})
 
-#######################
-
+# ######################
 find_package(gflags REQUIRED)
 find_package(Threads REQUIRED)
